renamed utils folder to lib
