+++ package.json
@@ -11,7 +11,10 @@
   },
   "dependencies": {
     "@lexical/react": "^0.4.1",
+    "@mdx-js/loader": "^2.2.1",
+    "@mdx-js/react": "^2.2.1",
     "@next-auth/prisma-adapter": "^1.0.4",
+    "@next/mdx": "^13.1.5",
     "@prisma/client": "^4.3.1",
     "@supabase/supabase-js": "^2.0.4",
     "@trpc/client": "^9.27.2",
@@ -22,6 +25,7 @@
     "lexical": "^0.4.1",
     "next": "12.2.5",
     "next-auth": "^4.18.4",
+    "next-mdx-remote": "^4.3.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-query": "^3.39.2",
