+++ src/lib/trpc.ts
@@ -1,4 +1,4 @@
-// src/utils/trpc.ts
+// src/lib/trpc.ts
 import type { AppRouter } from "../server/router";
 import { createReactQueryHooks } from "@trpc/react";
 import type { inferProcedureOutput, inferProcedureInput } from "@trpc/server";
