+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import type { NextPage } from "next";
 import Head from "next/head";
 import React from "react";
 import { AboutItem, Contacts } from "../components";
-import { aboutData } from "../utils/AboutData";
+import { aboutData } from "../lib/AboutData";
 
 const Home: NextPage = () => {
   return (
